@@ -15,54 +15,60 @@ using System.Windows.Media.Imaging;
 using System.Windows.Navigation;
 using System.Windows.Shapes;
 
-namespace TimeSpanDemo
+namespace TimeSpanDemo;
+
+public partial class MainWindow : Window, INotifyPropertyChanged
 {
-    public partial class MainWindow : Window, INotifyPropertyChanged
+    private TimeSpan? time1;
+    public TimeSpan? Time1
     {
-        private TimeSpan? time1;
-        public TimeSpan? Time1
-        {
-            get => time1;
-            set => SetField(ref time1, value);
-        }
-
-        private TimeSpan? _time2;
-        public TimeSpan? Time2
-        {
-            get => _time2;
-            set => SetField(ref _time2, value);
-        }
-
-        private TimeSpan? _time3;
-        public TimeSpan? Time3
-        {
-            get => _time3;
-            set => SetField(ref _time3, value);
-        }
-
-        private TimeSpan? _time4;
-        public TimeSpan? Time4
-        {
-            get => _time4;
-            set => SetField(ref _time4, value);
-        }
-
-        public MainWindow()
-        {
-            InitializeComponent();
-            DataContext = this;
-        }
-
-        #region INotifyPropertyChanged
-        public event PropertyChangedEventHandler PropertyChanged;
-        protected void OnPropertyChanged(string propertyName) => PropertyChanged?.Invoke(this, new PropertyChangedEventArgs(propertyName));
-        protected bool SetField<T>(ref T field, T value, [CallerMemberName] string propertyName = null)
-        {
-            if (EqualityComparer<T>.Default.Equals(field, value)) return false;
-            field = value;
-            OnPropertyChanged(propertyName);
-            return true;
-        }
-        #endregion
+        get => time1;
+        set => SetField(ref time1, value);
     }
+
+    private TimeSpan? _time2;
+    public TimeSpan? Time2
+    {
+        get => _time2;
+        set => SetField(ref _time2, value);
+    }
+
+    private TimeSpan? _time3;
+    public TimeSpan? Time3
+    {
+        get => _time3;
+        set => SetField(ref _time3, value);
+    }
+
+    private TimeSpan? _time4;
+    public TimeSpan? Time4
+    {
+        get => _time4;
+        set => SetField(ref _time4, value);
+    }
+
+    private TimeSpan? _time5;
+    public TimeSpan? Time5
+    {
+        get => _time5;
+        set => SetField(ref _time5, value);
+    }
+
+    public MainWindow()
+    {
+        InitializeComponent();
+        DataContext = this;
+    }
+
+    #region INotifyPropertyChanged
+    public event PropertyChangedEventHandler PropertyChanged;
+    protected void OnPropertyChanged(string propertyName) => PropertyChanged?.Invoke(this, new PropertyChangedEventArgs(propertyName));
+    protected bool SetField<T>(ref T field, T value, [CallerMemberName] string propertyName = null)
+    {
+        if (EqualityComparer<T>.Default.Equals(field, value)) return false;
+        field = value;
+        OnPropertyChanged(propertyName);
+        return true;
+    }
+    #endregion
 }
